+++ openrec/postprocess/ar_postprocess.py
@@ -23,7 +23,7 @@ class ARLabelDecode(BaseRecLabelDecode):
         if isinstance(preds, list):
             preds = preds[-1]
         if isinstance(preds, torch.Tensor):
-            preds = preds.detach().cpu().numpy()
+            preds = preds.detach().cpu().float().numpy()
         preds_idx = preds.argmax(axis=2)
         preds_prob = preds.max(axis=2)
         text = self.decode(preds_idx, preds_prob)
